+++ src/FileManager/File.php
@@ -2,8 +2,6 @@
 
 namespace Boscho87\ChangelogChecker\FileManager;
 
-use Boscho87\ChangelogChecker\Exception\FileNotFoundException;
-
 /**
  * Class File
  */
@@ -21,27 +19,10 @@ class File implements FileInterface
      */
     public function __construct(string $path)
     {
-        try {
-            $path = realpath($path);
-            $this->filePath = $path;
-            $this->content = file_get_contents($path);
-            $handle = @fopen($path, "r");
-            if ($handle) {
-                while (($buffer = fgets($handle)) !== false) {
-                    $this->lines[] = trim($buffer);
-                }
-                if (!feof($handle)) {
-                    echo "Fehler: unerwarteter fgets() Fehlschlag\n";
-                }
-                fclose($handle);
-            }
-        } catch (\Throwable $throwable) {
-            throw new FileNotFoundException(sprintf(
-                'File %s not found Error:{%s}',
-                $path,
-                $throwable->getMessage()
-            ));
-        }
+        $path = realpath($path);
+        $this->filePath = $path;
+        $content = file_get_contents($path);
+        $this->setNewContent($content);
     }
 
     public function getContents(): string
@@ -49,19 +30,14 @@ class File implements FileInterface
         return $this->content;
     }
 
-    public function getLines(): array
-    {
-        return $this->lines;
-    }
-
     public function getLine(int $number): string
     {
-        return $this->lines[$number];
+        return $this->lines[$number] ?? '';
     }
 
     public function current()
     {
-        return $this->lines[$this->currentLine];
+        return $this->lines[$this->currentLine] ?? null;
     }
 
     public function next()
@@ -84,28 +60,36 @@ class File implements FileInterface
         $this->currentLine = 0;
     }
 
+    /**
+     * @return int File line number is not the array index of line!
+     */
     public function lineNumber(): int
     {
         return $this->key() + 1;
     }
 
-    public function setLine(string $content, int $key)
+    public function setLine(string $content, int $key = null): void
     {
-        $this->lines[$key] = $content;
+        $line = $key ?? $this->lineNumber() - 1;
+        $this->lines[$line] = $content;
         $this->content = implode(PHP_EOL, $this->lines);
     }
 
-    public function write(): void
+    public function write(string $filePath = null): void
     {
         $lastLine = count($this->lines) - 1;
         $lastLineContent = $this->lines[$lastLine];
         if (!empty($lastLineContent)) {
-            $this->setLine('', $lastLine);
+            $this->setLine('', $lastLine + 1);
+        }
+        if ($filePath) {
+            file_put_contents($filePath, $this->content);
+            return;
         }
         file_put_contents($this->filePath, $this->content);
     }
 
-    public function writeBackup(): void
+    public function writeBackup(): string
     {
         $pos = strrpos($this->filePath, '/');
         $filename = substr($this->filePath, $pos + 1);
@@ -119,9 +103,13 @@ class File implements FileInterface
 
         $this->removeOldBackups();
         file_put_contents($filePath, $this->getContents());
+        return $filePath;
     }
 
 
+    /**
+     * @codeCoverageIgnore
+     */
     protected function removeOldBackups(): void
     {
         $backupFiles = glob(sprintf('%s/%s.*', getcwd(), $this->backupFileSuffix));
@@ -134,5 +122,25 @@ class File implements FileInterface
             array_pop($backupFiles);
             unlink($backupFiles[$index]);
         }
+    }
+
+    public function setNewContent(string $contents)
+    {
+        $this->content = $contents;
+        $this->lines = explode(PHP_EOL, $contents);
+    }
+
+    public function includeLinesAfter(array $lines, int $key = null)
+    {
+        $fileLines = $this->lines;
+        $line = $key ?? $this->lineNumber();
+        array_splice($fileLines, $line, 0, $lines);
+        $this->lines = $fileLines;
+        $this->content = implode(PHP_EOL, $this->lines);
+    }
+
+    public function getHash(): string
+    {
+        return md5($this->getContents());
     }
 }