+++ src/FileManager/FileInterface.php
@@ -9,11 +9,13 @@ interface FileInterface extends \Iterator
 {
     public function getContents(): string;
 
-    public function getLines(): array;
-
     public function getLine(int $number): string;
 
     public function lineNumber(): int;
 
-    public function setLine(string $content, int $key);
+    public function setLine(string $content, int $key = null);
+
+    public function setNewContent(string $contents);
+
+    public function includeLinesAfter(array $lines, int $key = null);
 }