+++ CHANGELOG.md
@@ -4,19 +4,49 @@ to [Semantic Versioning](http://semver.org/).
 # Changelog
 
 ## [Unreleased]
+### Added
+- Added test Cases
+- Updated Create Release Command
+- Refactored the Checkers after test Updates
 
-- Added Changelog Checker
-- Added 
+## [1.2.0] - 2021-07-03
+### Added
+- Added Create Release Command (Alpha)
 
-## [0.0.4] - 2021-06-27
+## [1.1.0] - 2021-06-29
+### Added
+- DefaultChecker
+- Implement fix logic in IncreaseChecker
 
+### Changed
+- Rename Increased Checker to ChangeChecker
+
+## [1.0.0] - 2021-06-28
+### Added
+- Added Changelog Checkers## [0.0.4] - 2021-06-27
 ### Added
 - Created Repository
 - Basic setup and Configuration
 - Implement Ascending version checker
 - Create Backup before manipulate the file
 
+## [0.0.3] - 2021-06-27
+### Fixed
+- Fixed some minor Bugs
 
+## [0.0.2] - 2021-06-27
+### Fixed
+- Fixed some minor Bugs
 
-[Unreleased]: 
-[0.0.4]: https://github.com/boscho87/changelog-checker/releases/tag/0.0.4
+## [0.0.1] - 2021-06-27
+### Added
+- Create Project
+
+[Unreleased]: https://github.com/boscho87/changelog-checker/compare/1.2.0...master
+[1.2.0]: https://github.com/boscho87/changelog-checker/compare/1.1.0...1.2.0
+[1.1.0]: https://github.com/boscho87/changelog-checker/compare/1.0.0...1.1.0
+[1.0.0]: https://github.com/boscho87/changelog-checker/compare/0.0.4...1.0.0
+[0.0.4]: https://github.com/boscho87/changelog-checker/compare/0.0.3...0.0.4
+[0.0.3]: https://github.com/boscho87/changelog-checker/compare/0.0.2...0.0.3
+[0.0.2]: https://github.com/boscho87/changelog-checker/compare/0.0.1...0.0.2
+[0.0.1]: https://github.com/boscho87/changelog-checker/releases/tag/0.0.1